Rename `Types` folder to lower-case `types` for consistency
