+++ src/comparison/lambdas/rerunComparisonLambda.integration.test.ts
@@ -15,8 +15,8 @@ import DynamoGateway from "../DynamoGateway/DynamoGateway"
 import createDynamoDbConfig from "../lib/createDynamoDbConfig"
 import createS3Config from "../lib/createS3Config"
 import InvokeCompareLambda from "../lib/InvokeCompareLambda"
-import type { ComparisonLog } from "../Types"
-import { isError } from "../Types"
+import type { ComparisonLog } from "../types"
+import { isError } from "../types"
 import compareLambda from "./compareLambda"
 import rerunComparisonLambda from "./rerunComparisonLambda"
