+++ src/comparison/lib/getFailedComparisonResults.integration.test.ts
@@ -2,8 +2,8 @@ import MockDynamo from "tests/helpers/MockDynamo"
 import "tests/helpers/setEnvironmentVariables"
 import dynamoDbTableConfig from "tests/helpers/testDynamoDbTableConfig"
 import DynamoGateway from "../DynamoGateway/DynamoGateway"
-import type { ComparisonLog } from "../Types"
-import { isError } from "../Types"
+import type { ComparisonLog } from "../types"
+import { isError } from "../types"
 import createDynamoDbConfig from "./createDynamoDbConfig"
 import getFailedComparisonResults from "./getFailedComparisonResults"
