+++ src/comparison/lib/logInDynamoDb.integration.test.ts
@@ -4,8 +4,8 @@ import "tests/helpers/setEnvironmentVariables"
 import MockDynamo from "../../../tests/helpers/MockDynamo"
 import dynamoDbTableConfig from "../../../tests/helpers/testDynamoDbTableConfig"
 import DynamoGateway from "../DynamoGateway/DynamoGateway"
-import type { ComparisonLog } from "../Types"
-import { isError } from "../Types"
+import type { ComparisonLog } from "../types"
+import { isError } from "../types"
 import createDynamoDbConfig from "./createDynamoDbConfig"
 import logInDynamoDb from "./logInDynamoDb"
