+++ src/comparison/types/CompareLambdaEvent.ts
