+++ src/comparison/types/ComparisonLog.ts
