+++ src/comparison/types/DynamoDbConfig.ts
