+++ src/comparison/types/Result.ts
