+++ src/comparison/types/index.ts
