+++ tests/helpers/processTestFile.ts
@@ -1,7 +1,7 @@
 import fs from "fs"
 import orderBy from "lodash.orderby"
 import path from "path"
-import type { ImportedComparison } from "../../src/comparison/Types/ImportedComparison"
+import type { ImportedComparison } from "../../src/comparison/types/ImportedComparison"
 import type { Trigger } from "../../src/types/Trigger"
 import type { TriggerCode } from "../../src/types/TriggerCode"
